style : split code files and fit norm
